backend/requirements.txt
# backend/requirements.txt
flask
flask-cors
psycopg2
pandas
sqlalchemy
requests
python-dotenv